Deduplicate Cloudflare AI helpers
