+++ app/utils/cloudflare-ai-transcription.server.ts
@@ -4,14 +4,8 @@ type WhisperTranscriptionResponse = {
 }
 
 import { Buffer } from 'node:buffer'
+import { getWorkersAiRunUrl } from './cloudflare-ai-utils.server.ts'
 import { getEnv } from './env.server.ts'
-
-function getWorkersAiRunUrl(model: string) {
-	// Cloudflare's REST route expects the model as path segments (with `/`), so do
-	// not URL-encode the model string (encoding can yield "No route for that URI").
-	const env = getEnv()
-	return `https://gateway.ai.cloudflare.com/v1/${env.CLOUDFLARE_ACCOUNT_ID}/${env.CLOUDFLARE_AI_GATEWAY_ID}/workers-ai/${model}`
-}
 
 export async function transcribeMp3WithWorkersAi({
 	mp3,
